### 2023-01-27 #5110 @editorialbot
[PRE REVIEW]: timeseriesflattener: A Python package for summarising features from (medical) time series
**Submitting author:** @MartinBernstorff (<a href="http://orcid.org/0000-0002-0234-5390">Martin Bernstorff</a>)
**Repository:** https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener
**Branch with paper.md** (empty if default branch):
**Version:** 0.22.1
