### 2023-01-27 #5110 @arfon
> Author (Jakob Damgaard) is missing affiliation (Theoj::Error)

Looks like an author is missing an affiliation here.
