### 2023-01-27 #5110 @MartinBernstorff
> @MartinBernstorff - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
>
> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

Ah, sorry to hear it! Any indication of the current size of the waitlist?
