### 2023-01-30 4ed8f1d
fix:(deps): update dask requirement
Updates the requirements on [dask](https://github.com/dask/dask) to permit the latest version.
- [Release notes](https://github.com/dask/dask/releases)
- [Changelog](https://github.com/dask/dask/blob/main/docs/release-procedure.md)
- [Commits](https://github.com/dask/dask/compare/2022.9.0...2023.1.1)

---
updated-dependencies:
- dependency-name: dask
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
