+++ pyproject.toml
@@ -16,7 +16,7 @@ dill = ">=0.3.0,<0.3.7"
 srsly = ">=2.4.4,<2.4.6"
 SQLAlchemy = ">=1.4.41, <1.5.42"
 pyodbc = ">=4.0.34, <4.0.36"
-dask = ">=2022.9.0,<2022.13.0"
+dask = ">=2022.9.0,<2023.2.0"
 pydantic = ">=1.9.0, <1.10.0"
 psutil = ">=5.9.1, <6.0.0"
 pandas = ">=1.4.0,<1.6.0"
