### 2023-01-30 8f09e23
fix:(deps): update pyarrow requirement
Updates the requirements on [pyarrow](https://github.com/apache/arrow) to permit the latest version.
- [Release notes](https://github.com/apache/arrow/releases)
- [Commits](https://github.com/apache/arrow/compare/go/v9.0.0...apache-arrow-11.0.0)

---
updated-dependencies:
- dependency-name: pyarrow
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
