### 2023-01-30 7e903aa
added loader_kwargs to all spec object classes and wrote a test for group combinations
