+++ tests/test_timeseriesflattener/test_feature_spec_objects.py
@@ -132,6 +132,7 @@ def test_skip_one_if_no_need_to_process():
     assert len(created_combinations) == 4
 
 
+
 def test_resolve_multiple_fn_to_str():
     """Test that resolve_multiple_fn is converted to str correctly."""
     pred_spec_batch = PredictorGroupSpec(
@@ -208,4 +209,3 @@ def test_predictorgroupspec_combinations_loader_kwargs(PredictorGroupSpec):
 
     assert olanzapine_100_rows == combinations[0].values_df
     assert clozapine_100_rows == combinations[1].values_df
-    
