### 2023-02-01 #5110 @MartinBernstorff
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/4024800881?check_suite_focus=true).

This is compiling correctly in the source repo now. Let me know if there's a way I can trigger this to recompile :-)
