### 2023-02-01 #5110 @danielskatz
Use the command `@editorialbot generate pdf`

editorialbot commands must be the first thing in a new comment
