### 2023-02-03 2986203
feat: TextSpec and option for additional feature name .get_col_str
