+++ tests/test_timeseriesflattener/test_feature_spec_objects.py
@@ -12,6 +12,7 @@ from timeseriesflattener.feature_spec_objects import (
     PredictorGroupSpec,
     PredictorSpec,
     TemporalSpec,
+    TextPredictorSpec,
     _AnySpec,
     check_that_col_names_in_kwargs_exist_in_df,
     generate_docstring_from_attributes,
@@ -150,6 +151,7 @@ def test_resolve_multiple_fn_to_str():
         TemporalSpec,
         PredictorSpec,
         PredictorGroupSpec,
+        TextPredictorSpec,
         OutcomeSpec,
         OutcomeGroupSpec,
     ],
