### 2023-02-03 a3cafe9
feat: handle text features and features with multiple columns
