+++ src/timeseriesflattener/flattened_dataset.py
@@ -9,7 +9,7 @@ import time
 from collections.abc import Callable
 from datetime import timedelta
 from multiprocessing import Pool
-from typing import Optional, Union
+from typing import List, Optional, Union
 
 import coloredlogs
 import numpy as np
@@ -26,6 +26,7 @@ from timeseriesflattener.feature_spec_objects import (
     PredictorSpec,
     StaticSpec,
     TemporalSpec,
+    TextPredictorSpec,
     _AnySpec,
 )
 from timeseriesflattener.flattened_ds_validator import ValidateInitFlattenedDataset
@@ -304,6 +305,24 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
         )
 
     @staticmethod
+    def _embed_text_values(
+        df: pd.DataFrame,
+        embedding_fn: Callable,
+        dim_reduction_fn: Optional[Callable],
+    ) -> pd.DataFrame:
+        """Embeds text values using the embedding_fn and optionally reduces the
+        dimensionality using dim_reduction_fn. Assumes that the text is stored
+        in the column "value"."""
+        embedding = embedding_fn(df["value"])
+        if dim_reduction_fn:
+            embedding = dim_reduction_fn(embedding)
+
+        df = df.drop("value", axis=1)
+        # make multiindex with embedding as 'value'
+        df = pd.concat([df, embedding], axis=1, keys=["df", "value"])
+        return df
+
+    @staticmethod
     def _flatten_temporal_values_to_df(  # noqa pylint: disable=too-many-locals
         prediction_times_with_uuid_df: DataFrame,
         output_spec: _AnySpec,
@@ -381,25 +400,121 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
             val_timestamp_col_name=timestamp_val_col_name,
         )
 
+        # handle embedding and dimensionality reduction if text predictor
+        if isinstance(output_spec, TextPredictorSpec):
+            df = TimeseriesFlattener._embed_text_values(
+                df=df,
+                embedding_fn=output_spec.embedding_fn,
+                dim_reduction_fn=output_spec.dim_reduction_fn,
+            )
+
         # If resolve_multiple generates empty values,
         # e.g. when there is only one prediction_time within look_ahead window for slope calculation,
         # replace with NaN
 
         # Rename column
-        df.rename(columns={"value": output_spec.get_col_str()}, inplace=True)
+        df = TimeseriesFlattener._rename_value_column(df=df, output_spec=output_spec)
 
         # Find value_cols and add fallback to them
-        df[output_spec.get_col_str()] = df[output_spec.get_col_str()].fillna(
-            output_spec.fallback,
-            inplace=False,
+        value_col_str_name = TimeseriesFlattener._get_value_col_str_name(
+            df=df, output_spec=output_spec
+        )
+        df = TimeseriesFlattener._add_fallback_to_value_cols(
+            df=df,
+            output_spec=output_spec,
         )
 
+        # check if multiindex and flatten
+        df = TimeseriesFlattener._flatten_multiindex(df)
         if verbose:
             log.info(
                 f"Returning {df.shape[0]} rows of flattened dataframe for {output_spec.get_col_str()}",
             )
 
-        return df[[pred_time_uuid_col_name, output_spec.get_col_str()]]
+        return df[value_col_str_name + [pred_time_uuid_col_name]]
+
+    @staticmethod
+    def _get_value_col_str_name(df: pd.DataFrame, output_spec=TemporalSpec) -> List[str]:
+        """Returns the name of the value column in df. If df has a multiindex,
+        returns a list of all column names in the 'value' multiindex.
+        
+        Args:
+            df (pd.DataFrame): Dataframe to get value column name from.
[…]
