### 2023-02-03 7230a96
tests: fixed tests to work with new features
