+++ tests/test_timeseriesflattener/test_embedding_functions.py
@@ -1,3 +1,4 @@
+"""Tests for the text embedding functions"""
 import pytest
 
 from timeseriesflattener.testing.text_embedding_functions import bow_test_embedding
