### 2023-02-03 f4ce9a2
feat: example text embedding functions
