### 2023-02-03 51b0796
build: add extra "text" dependencies
