+++ pyproject.toml
@@ -51,6 +51,11 @@ sphinx_design = ">=0.3.0,<0.3.1"
 autodoc_pydantic = ">=1.1.0,<1.9.0"
 
 
+[tool.poetry.group.text.dependencies]
+transformers = "^4.26.0"
+torch = "^1.13.1"
+sentence-transformers = "^2.2.2"
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
