### 2023-02-05 6e30b77
chore: remove `dim_reduction_fn` option in `TextPredictorSpec`
