### 2023-02-05 6416ed6
refactor: move multiindex handling to separate class
