+++ src/timeseriesflattener/flattened_dataset.py
@@ -30,7 +30,9 @@ from timeseriesflattener.feature_spec_objects import (
     _AnySpec,
 )
 from timeseriesflattener.flattened_ds_validator import ValidateInitFlattenedDataset
+from timeseriesflattener.multi_index_handling import MultiIndexHandler
 from timeseriesflattener.resolve_multiple_functions import resolve_multiple_fns
+from timeseriesflattener.text_embedding_functions import embed_text_column
 from timeseriesflattener.utils import print_df_dimensions_diff
 
 ProgressBar().register()
@@ -305,24 +307,6 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
         )
 
     @staticmethod
-    def _embed_text_values(
-        df: pd.DataFrame,
-        embedding_fn: Callable,
-        dim_reduction_fn: Optional[Callable],
-    ) -> pd.DataFrame:
-        """Embeds text values using the embedding_fn and optionally reduces the
-        dimensionality using dim_reduction_fn. Assumes that the text is stored
-        in the column "value"."""
-        embedding = embedding_fn(df["value"])
-        if dim_reduction_fn:
-            embedding = dim_reduction_fn(embedding)
-
-        df = df.drop("value", axis=1)
-        # make multiindex with embedding as 'value'
-        df = pd.concat([df, embedding], axis=1, keys=["df", "value"])
-        return df
-
-    @staticmethod
     def _flatten_temporal_values_to_df(  # noqa pylint: disable=too-many-locals
         prediction_times_with_uuid_df: DataFrame,
         output_spec: _AnySpec,
@@ -402,10 +386,8 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
 
         # handle embedding and dimensionality reduction if text predictor
         if isinstance(output_spec, TextPredictorSpec):
-            df = TimeseriesFlattener._embed_text_values(
-                df=df,
-                embedding_fn=output_spec.embedding_fn,
-                dim_reduction_fn=output_spec.dim_reduction_fn,
+            df = embed_text_column(
+                df=df, text_col_name="value", embedding_fn=output_spec.embedding_fn
             )
 
         # If resolve_multiple generates empty values,
@@ -413,114 +395,26 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
         # replace with NaN
 
         # Rename column
-        df = TimeseriesFlattener._rename_value_column(df=df, output_spec=output_spec)
+        df = MultiIndexHandler.rename_value_column(df=df, output_spec=output_spec)
 
         # Find value_cols and add fallback to them
-        value_col_str_name = TimeseriesFlattener._get_value_col_str_name(
+        value_col_str_name = MultiIndexHandler.get_value_col_str_name(
             df=df,
             output_spec=output_spec,
         )
-        df = TimeseriesFlattener._add_fallback_to_value_cols(
+        df = MultiIndexHandler.add_fallback_to_value_cols(
             df=df,
             output_spec=output_spec,
         )
 
         # check if multiindex and flatten
-        df = TimeseriesFlattener._flatten_multiindex(df)
+        df = MultiIndexHandler.flatten_multiindex(df)
         if verbose:
             log.info(
                 f"Returning {df.shape[0]} rows of flattened dataframe for {output_spec.get_col_str()}",
             )
 
         return df[value_col_str_name + [pred_time_uuid_col_name]]
-
-    @staticmethod
-    def _get_value_col_str_name(
-        df: pd.DataFrame,
-        output_spec=TemporalSpec,
-    ) -> List[str]:
-        """Returns the name of the value column in df. If df has a multiindex,
-        returns a list of all column names in the 'value' multiindex.
-
-        Args:
-            df (pd.DataFrame): Dataframe to get value column name from.
[…]
