### 2023-02-05 6533578
tests: remove `dim_reduction_fn` from tests
