### 2023-02-05 7aa48a5
feat: add `embedding_fn_kwargs` to `TextPredictorSpec`
