### 2023-02-06 184c857
fix:(deps): update deepchecks requirement
Updates the requirements on [deepchecks](https://github.com/deepchecks/deepchecks) to permit the latest version.
- [Release notes](https://github.com/deepchecks/deepchecks/releases)
- [Commits](https://github.com/deepchecks/deepchecks/compare/0.8.0...0.11.1)

---
updated-dependencies:
- dependency-name: deepchecks
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
