+++ pyproject.toml
@@ -11,7 +11,7 @@ scipy = ">=1.8.0,<1.9.4"
 wasabi = ">=0.9.1,<1.1.1"
 scikit-learn = ">=1.1.2, <1.1.3"
 wandb = ">=0.12.0,<0.13.8"
-deepchecks = ">=0.8.0,<0.11.0"
+deepchecks = ">=0.8.0,<0.12.0"
 dill = ">=0.3.0,<0.3.7"
 srsly = ">=2.4.4,<2.4.6"
 SQLAlchemy = ">=1.4.41, <1.5.42"
