+++ pyproject.toml
@@ -9,7 +9,7 @@ authors = ["Martin Bernstorff", "Kenneth Enevoldsen", "Jakob Grøhn Damgaard", "
 python = ">=3.9, <3.11"
 scipy = ">=1.8.0,<1.9.4"
 wasabi = ">=0.9.1,<1.1.1"
-scikit-learn = ">=1.1.2, <1.1.3"
+scikit-learn = ">=1.1.2,<1.2.2"
 wandb = ">=0.12.0,<0.13.8"
 deepchecks = ">=0.8.0,<0.11.0"
 dill = ">=0.3.0,<0.3.7"
