### 2023-02-06 c10565d
fix: correct col str name generation for text specs when column names are ints (e.g. HF embeddings)
