### 2023-02-06 6a1cbc7
tests: add TextPredictorSpec to compute specs tests
