### 2023-02-06 6c10d3f
feat: add sklearn embedding function
