+++ tests/test_timeseriesflattener/test_flattened_dataset/test_flattened_dataset.py
@@ -108,7 +108,7 @@ def test_compute_specs(
         fallback=np.nan,
         embedding_fn=sentence_transformers_embedding,
         embedding_fn_kwargs={
-            "model_name": "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2"
+            "model_name": "sentence-transformers/paraphrase-multilingual-MiniLM-L12-v2",
         },
     )
 
