+++ tests/test_data/models/create_bow_and_pca_model.py
@@ -5,7 +5,6 @@ from pathlib import Path
 from typing import Sequence
 
 import numpy as np
-import pandas as pd
 from sklearn.decomposition import PCA
 from sklearn.feature_extraction.text import CountVectorizer
 
