### 2023-02-06 d492a40
docs: add textpredictorspec docstring
