+++ src/timeseriesflattener/feature_spec_objects.py
@@ -325,7 +325,6 @@ class StaticSpec(_AnySpec):
 
 class TemporalSpec(_AnySpec):
     """The minimum specification required for collapsing a temporal
-
         feature, whether looking ahead or behind. Mostly used for inheritance below.
 
     Fields:
@@ -355,7 +354,7 @@ class TemporalSpec(_AnySpec):
         interval_days (Union[int, float]):
             How far to look in the given direction (ahead for outcomes,
             behind for predictors)
-        resolve_multiple_fn (Callable):
+        resolve_multiple_fn (Union[Callable, str]):
             A function used for resolving multiple values within the
             interval_days.
         key_for_resolve_multiple (Optional[str]):
@@ -370,7 +369,7 @@ class TemporalSpec(_AnySpec):
             If NaN is higher than this in the input dataframe during
             resolution, raise an error. Defaults to: 0.0.
         entity_id_col_name (str):
-    Col name for ids in the input dataframe. Defaults to: entity_id.
+            Col name for ids in the input dataframe. Defaults to: entity_id.
     """
 
     class Doc:
@@ -482,7 +481,7 @@ class PredictorSpec(TemporalSpec):
         interval_days (Union[int, float]):
             How far to look in the given direction (ahead for outcomes,
             behind for predictors)
-        resolve_multiple_fn (Callable):
+        resolve_multiple_fn (Union[Callable, str]):
             A function used for resolving multiple values within the
             interval_days.
         key_for_resolve_multiple (Optional[str]):
@@ -499,7 +498,7 @@ class PredictorSpec(TemporalSpec):
         entity_id_col_name (str):
             Col name for ids in the input dataframe. Defaults to: entity_id.
         lookbehind_days (Union[int, float]):
-    How far behind to look for values
+            How far behind to look for values
     """
 
     class Doc:
@@ -530,7 +529,61 @@ class PredictorSpec(TemporalSpec):
 
 
 class TextPredictorSpec(PredictorSpec):
-    """Specification for a text predictor, where the df has been resolved."""
+    """Specification for a text predictor, where the df has been resolved.
+
+    Fields:
+        values_loader (Optional[Callable]):
+            Loader for the df. Tries to resolve from the data_loaders registry,
+            then calls the function which should return a dataframe.
+        values_name (Optional[str]):
+            A string that maps to a key in a dictionary instantiated by
+            `split_df_and_register_to_dict`. Each key corresponds to a dataframe, which
+            is a subset of the df where the values_name == key.
+        loader_kwargs (Optional[dict]):
+            Optional kwargs passed onto the data loader.
+        values_df (Optional[DataFrame]):
+            Dataframe with the values.
+        feature_name (str):
+            The name of the feature. Used for column name generation, e.g.
+            <prefix>_<feature_name>.
+        prefix (str):
+            The prefix used for column name generation, e.g.
+            <prefix>_<feature_name>. Defaults to: pred.
+        input_col_name_override (Optional[str]):
+            An override for the input column name. If None, will  attempt
+            to infer it by looking for the only column that doesn't match id_col_name
+            or timestamp_col_name.
+        output_col_name_override (Optional[str]):
+            Override the generated column name after flattening the time series
+        interval_days (Union[int, float]):
+            How far to look in the given direction (ahead for outcomes,
+            behind for predictors)
+        resolve_multiple_fn (Union[str, Callable]):
+            A function used for resolving multiple values within the
+        interval_days, i.e. how to combine texts within the lookbehind window.
+        Defaults to: 'concatenate'. Other possible options are 'latest' and
+        'earliest'. Defaults to: concatenate.
+        key_for_resolve_multiple (Optional[str]):
[…]
