### 2023-02-06 aea9c81
test: add test for loader_kwargs with synth data
