+++ pyproject.toml
@@ -1,59 +1,64 @@
-[tool.poetry]
+[project
 name = "timeseriesflattener"
 version = "0.22.1"
 description = "A package for converting time series data from e.g. electronic health records into wide format data."
 readme = "README.md"
 authors = ["Martin Bernstorff", "Kenneth Enevoldsen", "Jakob Grøhn Damgaard", "Frida Hæstrup", "Lasse Hansen"]
 
-[tool.poetry.dependencies]
-python = ">=3.9, <3.11"
-scipy = ">=1.8.0,<1.9.4"
-wasabi = ">=0.9.1,<1.1.2"
-scikit-learn = ">=1.1.2,<1.2.2"
-wandb = ">=0.12.0,<0.13.8"
-deepchecks = ">=0.8.0,<0.12.0"
-dill = ">=0.3.0,<0.3.7"
-srsly = ">=2.4.4,<2.4.6"
-SQLAlchemy = ">=1.4.41, <1.5.42"
-pyodbc = ">=4.0.34, <4.0.36"
-dask = ">=2022.9.0,<2023.2.0"
-pydantic = ">=1.9.0, <1.10.0"
-psutil = ">=5.9.1, <6.0.0"
-pandas = ">=1.4.0,<1.6.0"
-catalogue = ">=2.0.0, <2.1.0"
-numpy = ">=1.23.3,<1.24.2"
-pyarrow = ">=9.0.0,<11.1.0"
-psycopmlutils = ">=0.2.4,<0.4.0"
-protobuf = "<=3.20.3" # Other versions give errors with pytest, super weird!
-frozendict = ">=2.3.4,<2.4.0"
-skimpy = ">=0.0.7,<0.1.0"
-jupyter = ">=1.0.0,<1.1.0"
-coloredlogs = ">14.0.0,<15.1.0"
+requires-python = ">=3.8.0
 
-[tool.poetry.dev-dependencies]
-pytest = ">=7.1.3, <7.1.5"
-black = ">=22.8.0,<22.12.1"
-pre-commit = ">=2.20.0,<2.21.1"
-pytest-cov = ">=3.0.0,<4.0.1"
-flake8 = ">=5.0.0, <5.0.6"
-docformatter = ">=1.5.0, <1.5.2"
-mypy = ">=0.971,<0.992"
-pytest-xdist = ">=2.4.0,<3.1.1"
-pylint = ">=2.15.5,<2.16.0"
+dependencies = [
+    "python>=3.9, <3.11",
+    "scipy>=1.8.0,<1.9.4",
+    "wasabi>=0.9.1,<1.1.2",
+    "scikit-learn>=1.1.2,<1.2.2",
+    "wandb>=0.12.0,<0.13.8",
+    "deepchecks>=0.8.0,<0.12.0",
+    "dill>=0.3.0,<0.3.7",
+    "srsly>=2.4.4,<2.4.6",
+    "SQLAlchemy>=1.4.41, <1.5.42",
+    "pyodbc>=4.0.34, <4.0.36",
+    "dask>=2022.9.0,<2023.2.0",
+    "pydantic>=1.9.0, <1.10.0",
+    "psutil>=5.9.1, <6.0.0",
+    "pandas>=1.4.0,<1.6.0",
+    "catalogue>=2.0.0, <2.1.0",
+    "numpy>=1.23.3,<1.24.2",
+    "pyarrow>=9.0.0,<11.1.0",
+    "psycopmlutils>=0.2.4,<0.4.0",
+    "protobuf<=3.20.3", # Other versions give errors with pytest, super weird!
+    "frozendict>=2.3.4,<2.4.0",
+    "skimpy>=0.0.7,<0.1.0",
[…]
