+++ .github/actions/test/action.yml
@@ -1,16 +1,10 @@
-name: 'Run tests'
-description: 'Run tests'
+name: "Run tests"
+description: "Run tests"
 inputs:
   python-version:
     description: "Which Python version to run on"
     required: true
     default: 3.9
-  poetry-version:
-    description: "Which version of Poetry to install"
-    default: 1.1.15
-  cache-version:
-    description: "Cache version for invalidation"
-    default: 0.0.0
 
 runs:
   using: "composite"
@@ -20,32 +14,18 @@ runs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ inputs.python-version }}
+        cache: "pip"
+        cache-dependency-path: "**/pyproject.toml"
 
-    ### Caching
-    - name: Cache poetry and venv
-      id: cache-poetry-and-venv
-      uses: MartinBernstorff/cache-poetry-and-venv@latest
-      with:
-        poetry-version: ${{ inputs.poetry-version }}
-        cache-version: ${{ inputs.cache-version }}
-        python-version: ${{ inputs.python-version }}
-
-    # Installing
-    - name: Install Poetry
-      if: steps.cache-poetry-and-venv.outputs.poetry-cache-hit != 'true'
-      uses: snok/install-poetry@v1
-      with:
-        version: ${{ inputs.poetry-version }}
-        virtualenvs-create: true
-        virtualenvs-in-project: true
-
-    - name: Install Project
-      if: steps.cache-poetry-and-venv.outputs.venv-cache-hit != 'true'
+    - name: Install dependencies
       shell: bash
-      run: poetry install -vvv
-    
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e ".[dev]"
     ### Run tests
     - name: Run pytest
       shell: bash
       run: |
-        poetry run pytest --cov=src --cov-report term-missing --color=yes
+        set -o pipefail
+        ls -la
+        pytest --cov=src --cov-report term-missing --color=yes -n 2
