### 2023-02-07 0dfdcf3
chore: clean up more poetry stuff
