+++ .github/actions/test_tutorials/action.yml
@@ -15,14 +15,6 @@ runs:
         python-version: ${{ inputs.python-version }}
 
     # Installing
-    - name: Install Poetry
-      if: steps.cache-poetry-and-venv.outputs.poetry-cache-hit != 'true'
-      uses: snok/install-poetry@v1
-      with:
-        version: ${{ inputs.poetry-version }}
-        virtualenvs-create: true
-        virtualenvs-in-project: true
-
     - name: Install dependencies
       shell: bash
       run: |
