### 2023-02-07 21f1a7a
fix: authors must be a dict in new toml
