+++ pyproject.toml
@@ -3,7 +3,10 @@ name = "timeseriesflattener"
 version = "0.22.1"
 description = "A package for converting time series data from e.g. electronic health records into wide format data."
 readme = "README.md"
-authors = ["Martin Bernstorff", "Kenneth Enevoldsen", "Jakob Grøhn Damgaard", "Frida Hæstrup", "Lasse Hansen"]
+authors = [{name = "Martin Bernstorff", email = "martinbernstorff@gmail.com"},
+    {name = "Lasse Hansen", email = "lasseh0310@gmail.com"},
+    {name = "Jakob Grøhn Damgaard", email = ""},
+    {name = "Kenneth Enevoldsen"}]
 
 requires-python = ">=3.8.0"
 
