+++ pyproject.toml
@@ -40,7 +40,7 @@ dependencies = [
 dev = [
     "pytest>=7.1.3, <7.1.5",
     "black>=22.8.0,<22.12.1",
-    "premmit = >=2.20.0,<2.21.1",
+    "pre-commit>=2.20.0,<2.21.1",
     "pytestv = >=3.0.0,<4.0.1",
     "flake8>=5.0.0, <5.0.6",
     "docformatter>=1.5.0, <1.5.2",
