### 2023-02-07 2a9d649
build: remove python from deps
