+++ .github/actions/test/action.yml
@@ -22,6 +22,7 @@ runs:
       run: |
         python -m pip install --upgrade pip
         pip install -e ".[dev]"
+
     ### Run tests
     - name: Run pytest
       shell: bash
