### 2023-02-07 11722ff
build: add env to jupyter tutorials
