### 2023-02-07 fe2247f
build: remove poetry (#110)
* init: first stab

* chore: clean up more poetry stuff

* build: typo in toml

* chore: typo in toml

* fix: authors must be a dict in new toml

* build: ad bokajgd email

* build: typo in toml

* fix: typos in toml

* build: remove python from deps

* build: remove link from toml in case that's the cause of the error

* build: add python-version to env

* build: add env to jupyter tutorials
