+++ .github/workflows/documentation.yml
@@ -6,62 +6,44 @@ name: Documentation
 on:
   push:
     branches:
-    - main
-
+      - main
 
 env:
   cache-version: 0.0.4
-  poetry-version: 1.1.15
   python-version: 3.9
 
 jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-    - name: Checkout
-      uses: actions/checkout@v3
-      with:
-        fetch-depth: 0 # otherwise, you will failed to push refs to dest repo
+      - name: Checkout
+        uses: actions/checkout@v3
+        with:
+          fetch-depth: 0 # otherwise, you will failed to push refs to dest repo
 
-    ### Setup prerequisites
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: 3.9
+      ### Setup prerequisites
+      - name: Set up Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.9
 
-    ### Caching
-    - name: Cache poetry and venv
-      id: cache-poetry-and-venv
-      uses: MartinBernstorff/cache-poetry-and-venv@latest
-      with:
-        poetry-version: ${{ env.poetry-version }}
-        cache-version: ${{ env.cache-version }}
-        python-version: ${{ env.python-version }}
+      # Installing
+      - name: Install dependencies
+        shell: bash
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e ".[docs]"
 
-    # Installing
-    - name: Install Poetry
-      if: steps.cache-poetry-and-venv.outputs.poetry-cache-hit != 'true'
-      uses: snok/install-poetry@v1
-      with:
-        version: ${{ env.poetry-version }}
-        virtualenvs-create: true
-        virtualenvs-in-project: true
+      - name: Build documentation
+        run: |
+          mkdir gh-pages
+          touch gh-pages/.nojekyll
+          cd docs/
+          poetry run sphinx-build -b html . _build
+          cp -r _build/* ../gh-pages/
 
-    - name: Install Project
-      if: steps.cache-poetry-and-venv.outputs.venv-cache-hit != 'true'
-      shell: bash
-      run: poetry install -vvv
-
-    - name: Build documentation
-      run: |
-        mkdir gh-pages
-        touch gh-pages/.nojekyll
-        cd docs/
-        poetry run sphinx-build -b html . _build
-        cp -r _build/* ../gh-pages/
-
-    - name: Deploy documentation
-      uses: JamesIves/github-pages-deploy-action@4.1.4
-      with:
-        branch: gh-pages
-        folder: gh-pages
+      - name: Deploy documentation
+        uses: JamesIves/github-pages-deploy-action@4.1.4
+        with:
+          branch: gh-pages
+          folder: gh-pages
