+++ .github/workflows/documentation.yml
@@ -39,7 +39,7 @@ jobs:
           mkdir gh-pages
           touch gh-pages/.nojekyll
           cd docs/
-          poetry run sphinx-build -b html . _build
+          sphinx-build -b html . _build
           cp -r _build/* ../gh-pages/
 
       - name: Deploy documentation
