### 2023-02-07 21aa6e1
build: move text dependencies to pyproject optional
