### 2023-02-07 d67355a
ci: update actions to install text dependencies when running tests
