+++ .github/actions/test_tutorials/action.yml
@@ -19,7 +19,7 @@ runs:
       shell: bash
       run: |
         python -m pip install --upgrade pip
-        pip install -e ".[dev]"
+        pip install -e ".[dev,text]"
 
     ### Run tests
     - name: Install jupyter
