+++ .github/actions/test/action.yml
@@ -21,7 +21,7 @@ runs:
       shell: bash
       run: |
         python -m pip install --upgrade pip
-        pip install -e ".[dev]"
+        pip install -e ".[dev,text]"
 
     ### Run tests
     - name: Run pytest
