### 2023-02-07 5bb6a02
docs: add text tutorial to docs
