+++ docs/tutorials.rst
@@ -11,3 +11,4 @@ locally.
 
    tutorials/01_basic.ipynb
    tutorials/02_advanced.ipynb
+   tutorial/03_text.ipynb
