### 2023-02-07 f23b968
build: add sklearn dependency
