+++ pyproject.toml
@@ -62,6 +62,7 @@ text = [
     "transformers>=4.26.0,<4.27.0",
     "torch>=1.12.1,<1.14.0",
     "sentence-transformers>=2.2.0,<2.3.0",
+    "scikit-learn>=1.1.2,<1.4.0",
 ]
 
 [build-system]
