### 2023-02-07 8ed8fa3
chore: remove extra character in docstring
