### 2023-02-07 5188626
ci: don't run tests in parallel
