### 2023-02-07 d519934
tests: check if huggingface models messing up tests
