### 2023-02-08 679a246
tests: rename synthetic text data
