### 2023-02-08 2451e67
ci: don't run huggingface tests in CI
