+++ .github/actions/test/action.yml
@@ -29,4 +29,4 @@ runs:
       run: |
         set -o pipefail
         ls -la
-        pytest --cov=src --cov-report term-missing --color=yes
+        pytest --skiphuggingface --cov=src --cov-report term-missing --color=yes
