### 2023-02-09 aeef039
fix: proper guard against neither `df` nor `output_spec` specified in `get_value_col_str_name`
