### 2023-02-09 c6bffd2
build: fix furo version
