+++ pyproject.toml
@@ -50,7 +50,7 @@ dev = [
 
 docs = [
     "sphinx>=5.3.0,<5.4.0",
-    "furo2022.12.7",
+    "furo==2022.12.7",
     "sphinx-copybutton>=0.5.1,<0.5.2",
     "sphinxext-opengraph>=0.7.3,<0.7.5",
     "myst-nb>=0.6.0,<1.17.0",
