### 2023-02-09 487daec
fix: move `embed_text_values` to use timeseriesflattener without text dependencies installed
