### 2023-02-09 0451bdb
build: update documentation build command
