+++ docs/tutorials.rst
@@ -11,4 +11,4 @@ locally.
 
    tutorials/01_basic.ipynb
    tutorials/02_advanced.ipynb
-   tutorial/03_text.ipynb
+   tutorials/03_text.ipynb
