+++ .github/workflows/main_test_and_release.yml
@@ -74,7 +74,7 @@ jobs:
           token: ${{ secrets.RELEASE_BOT }}
 
       - name: Python Semantic Release
-        uses: relekang/python-semantic-release@v7.32.0
+        uses: relekang/python-semantic-release@v7.33.1
         with:
           github_token: ${{ secrets.RELEASE_BOT }}
           # Remember to copy the tool.semantic_release section from pyproject.toml
