+++ pyproject.toml
@@ -25,7 +25,7 @@ dependencies = [
     "psutil>=5.9.1, <6.0.0",
     "pandas>=1.4.0,<1.6.0",
     "catalogue>=2.0.0, <2.1.0",
-    "numpy>=1.23.3,<1.24.2",
+    "numpy>=1.23.3,<1.24.3",
     "pyarrow>=9.0.0,<11.1.0",
     "psycopmlutils>=0.2.4,<0.4.0",
     "protobuf<=3.20.3", # Other versions give errors with pytest, super weird!
