+++ pyproject.toml
@@ -44,7 +44,7 @@ dev = [
     "flake8>=5.0.0, <5.0.6",
     "docformatter>=1.5.0, <1.5.2",
     "mypy>=0.971,<0.992",
-    "pytest-xdist>=2.4.0,<3.1.1",
+    "pytest-xdist>=2.4.0,<3.2.1",
     "pylint>=2.15.5,<2.16.0",
 ]
 
