+++ html/_sources/tutorials.rst.txt
@@ -1,14 +1,0 @@
-Tutorials
-================================
-
-To get started using the package, we recommend going through the tutorials in the order
-listed below. Each tutorial also a jupyter notebook which you can download and run
-locally.
-
-.. toctree::
-   :maxdepth: 3
-   :caption: Getting started
-
-   tutorials/01_basic.ipynb
-   tutorials/02_advanced.ipynb
-   tutorials/03_text.ipynb
