+++ html/_sources/timeseriesflattener.rst.txt
@@ -1,14 +1,0 @@
-Timeseriesflattener
---------------------------------------------------
-
-.. FlattenedDataset
-
-
-timeseriesflattener.flattened_dataset 
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: timeseriesflattener.flattened_dataset
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
