+++ .github/pull_request_template.md
@@ -1,7 +1,0 @@
-- [ ] I have battle-tested on Overtaci (RMAPPS1279)
-- [ ] I have assigned ranges (e.g. `>=0.1, <0.2`) to all new dependencies (allows dependabot to keep dependency ranges wide for better compatability)
-
-Fixes #[issue_nr_here].
-
-## Notes for reviewers
-Reviewers can skip X, but should pay attention to Y.
