+++ html/_sources/feature_specifications.rst.txt
@@ -1,14 +1,0 @@
-Feature specifications
---------------------------------------------------
-
-.. FlattenedDataset
-
-
-timeseriesflattener.feature_spec_objects
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: timeseriesflattener.feature_spec_objects
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
