+++ html/_sources/installation.rst.txt
@@ -1,10 +1,0 @@
-Installation
-==================
-To get started using timeseriesflattener install it using pip by running the following line in your terminal:
-
-.. code-block:: bash
-
-   pip install timeseriesflattener
-
-
-There is no discrepancy between the latest version installed using pip or the version on GitHub.
