+++ html/_sources/tutorials/03_text.ipynb.txt
@@ -1,1637 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "# Adding text features\n",
-    "\n",
-    "So far, the tutorials have dealt with _tabular_ data only. This tutorial will show you to make predictors out of text features, such as clinical notes, within `timeseriesflattener`.\n",
-    "\n",
-    "Specifically, this tutorial will cover:\n",
-    "\n",
-    "1. How to featurize text using Huggingface or sci-kit learn models.\n",
-    "2. How to use write your own text embedding function in `timeseriesflattener`.\n",
-    "\n",
-    "To use the features in this tutorial you'll need to install some extra dependencies. These can be installed by running:\n",
-    "```\n",
-    "pip install pytorch transformers sentence-transformer\n",
-    "```\n",
-    "or by installing `timeseriesflattener` with the text dependencies.\n",
-    "```\n",
-    "pip install timeseriesflattener\"[text]\"\n",
-    "```"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "metadata": {},
-   "source": [
-    "## The dataset\n",
-    "\n",
-    "To start out, let's load a synthetic dataset containing text. As with all other features, each row in the dataset needs an ID, a timestamp, and the feature value. "
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "from timeseriesflattener.testing.load_synth_data import load_synth_text"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 2,
-   "metadata": {},
-   "outputs": [
-    {
-     "data": {
-      "text/html": [
-       "<div>\n",
-       "<style scoped>\n",
-       "    .dataframe tbody tr th:only-of-type {\n",
-       "        vertical-align: middle;\n",
-       "    }\n",
-       "\n",
-       "    .dataframe tbody tr th {\n",
-       "        vertical-align: top;\n",
-       "    }\n",
-       "\n",
-       "    .dataframe thead th {\n",
-       "        text-align: right;\n",
-       "    }\n",
-       "</style>\n",
-       "<table border=\"1\" class=\"dataframe\">\n",
-       "  <thead>\n",
-       "    <tr style=\"text-align: right;\">\n",
-       "      <th></th>\n",
-       "      <th>entity_id</th>\n",
-       "      <th>timestamp</th>\n",
-       "      <th>text</th>\n",
-       "    </tr>\n",
-       "  </thead>\n",
-       "  <tbody>\n",
-       "    <tr>\n",
-       "      <th>0</th>\n",
-       "      <td>4647</td>\n",
-       "      <td>1967-07-19 00:22:00</td>\n",
-       "      <td>The patient went into a medically induced coma...</td>\n",
-       "    </tr>\n",
-       "    <tr>\n",
-       "      <th>1</th>\n",
-       "      <td>2007</td>\n",
-       "      <td>1966-11-25 02:02:00</td>\n",
-       "      <td>The patient is taken to the emergency departme...</td>\n",
-       "    </tr>\n",
-       "    <tr>\n",
-       "      <th>2</th>\n",
-       "      <td>5799</td>\n",
-       "      <td>1967-09-19 12:31:00</td>\n",
-       "      <td>The patient, described as a 7-month old son wh...</td>\n",
-       "    </tr>\n",
-       "    <tr>\n",
-       "      <th>3</th>\n",
-       "      <td>1319</td>\n",
-       "      <td>1969-07-21 23:16:00</td>\n",
[…]
