+++ .github/dependabot.yml
@@ -1,20 +1,0 @@
-# To get started with Dependabot version updates, you'll need to specify which
-# package ecosystems to update and where the package manifests are located.
-# Please see the documentation for all configuration options:
-# https://docs.github.com/github/administering-a-repository/configuration-options-for-dependency-updates
-
-version: 2
-updates:
-  - package-ecosystem: "pip" # See documentation for possible values
-    directory: "/" # Location of package manifests
-    schedule:
-      interval: "weekly"
-      day: "monday"
-      time: "01:00"
-      timezone: "Europe/Copenhagen"
-    open-pull-requests-limit: 5
-    commit-message:
-      prefix: "fix:"
-      include: "scope"
-    ignore:
-      - dependency-name: "protobuf"

[…]
