### 2023-02-17 0a434c3
ci: only move html folder to gh pages
