+++ .github/workflows/documentation.yml
@@ -32,14 +32,14 @@ jobs:
         shell: bash
         run: |
           python -m pip install --upgrade pip
-          pip install -e ".[docs,text]" 
+          pip install -e ".[docs,text]"
 
       - name: Build documentation
         run: |
           mkdir gh-pages
           touch gh-pages/.nojekyll
           make -C docs html
-          cp -r docs/_build/* gh-pages/
+          cp -r docs/_build/html/* gh-pages/
 
       - name: Deploy documentation
         uses: JamesIves/github-pages-deploy-action@4.1.4
