+++ _sources/tutorials/02_advanced.ipynb.txt
@@ -1,0 +1,708 @@
+{
+    "cells": [
+        {
+            "attachments": {},
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "# Advanced Tutorial"
+            ]
+        },
+        {
+            "attachments": {},
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "In the basic tutorial we covered how to add static features, predictors and outcomes.\n",
+                "In this tutorial, we'll expand on that, covering how to effectively add many features by:\n",
+                "1. Utilising data loaders in a data_loaders registry,\n",
+                "2. Populating a dictionary from a long format dataframe,\n",
+                "3. Creating feature combinations from specifcations,\n",
+                "4. Using caching, so you can iterate on your datasets without having to complete full computations every time\n"
+            ]
+        },
+        {
+            "attachments": {},
+            "cell_type": "markdown",
+            "metadata": {},
+            "source": [
+                "## Using data loaders\n",
+                "Until now, we've loaded our data first and then created combinations. But what if your data lies in an SQL database, and you don't want to save it to disk?\n",
+                "\n",
+                "Time to introduce feature loaders. All feature spec objects (PredictorSpec, OutcomeSpec and StaticSpec) can resolve from a loader function. The only requirement of that loader function is that it should return a values dataframe, which should contain an ID column, a timestamp column and value column. This means you can have loaders that load from REDIS, SQL databases, or just from disk. Whatever you prefer.\n",
+                "\n",
+                "This function is then called when you initialise a feature specification.\n",
+                "\n",
+                "This loader is specified in the values_loader key like so:\n"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": 1,
+            "metadata": {},
+            "outputs": [],
+            "source": [
+                "from skimpy import skim\n",
+                "from timeseriesflattener.testing.load_synth_data import load_synth_predictor_float\n",
+                "from timeseriesflattener.resolve_multiple_functions import mean\n",
+                "from timeseriesflattener.feature_spec_objects import PredictorSpec\n",
+                "from pprint import pprint\n",
+                "import numpy as np"
+            ]
+        },
+        {
+            "cell_type": "code",
+            "execution_count": 2,
+            "metadata": {},
+            "outputs": [
+                {
+                    "data": {
+                        "text/html": [
+                            "<div>\n",
+                            "<style scoped>\n",
+                            "    .dataframe tbody tr th:only-of-type {\n",
+                            "        vertical-align: middle;\n",
+                            "    }\n",
+                            "\n",
+                            "    .dataframe tbody tr th {\n",
+                            "        vertical-align: top;\n",
+                            "    }\n",
+                            "\n",
+                            "    .dataframe thead th {\n",
+                            "        text-align: right;\n",
+                            "    }\n",
+                            "</style>\n",
+                            "<table border=\"1\" class=\"dataframe\">\n",
+                            "  <thead>\n",
+                            "    <tr style=\"text-align: right;\">\n",
+                            "      <th></th>\n",
+                            "      <th>entity_id</th>\n",
+                            "      <th>timestamp</th>\n",
+                            "      <th>value</th>\n",
+                            "    </tr>\n",
+                            "  </thead>\n",
+                            "  <tbody>\n",
+                            "    <tr>\n",
+                            "      <th>0</th>\n",
+                            "      <td>9476</td>\n",
+                            "      <td>1969-03-05 08:08:00</td>\n",
[…]
