+++ pyproject.toml
@@ -11,7 +11,7 @@ authors = [{name = "Martin Bernstorff", email = "martinbernstorff@gmail.com"},
 requires-python = ">=3.8.0"
 
 dependencies = [
-    "scipy>=1.8.0,<1.9.4",
+    "scipy>=1.8.0,<1.10.2",
     "wasabi>=0.9.1,<1.1.2",
     "scikit-learn>=1.1.2,<1.2.2",
     "wandb>=0.12.0,<0.13.8",
