+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
     "deepchecks>=0.8.0,<0.12.0",
     "dill>=0.3.0,<0.3.7",
     "srsly>=2.4.4,<2.4.6",
-    "SQLAlchemy>=1.4.41, <1.5.42",
+    "SQLAlchemy>=1.4.41, <2.0.5",
     "pyodbc>=4.0.34, <4.0.36",
     "dask>=2022.9.0,<2023.2.0",
     "pydantic>=1.9.0, <1.10.0",
