### 2023-02-27 #5110 @mstimberg
:wave: @ejolly, @Ebedthan (and of course, your invitation still stands, @arianesasso),  would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
