### 2023-03-01 c417c30
chore: move codespell to pre-commit
