+++ .github/pull_request_template.md
@@ -1,4 +1,4 @@
-- [ ] I have assigned ranges (e.g. `>=0.1, <0.2`) to all new dependencies (allows dependabot to keep dependency ranges wide for better compatability)
+- [ ] I have assigned ranges (e.g. `>=0.1, <0.2`) to all new dependencies (allows dependabot to keep dependency ranges wide for better compatibility)
 
 Fixes #[issue_nr_here].
 
