+++ .pre-commit-config.yaml
@@ -1,6 +1,14 @@
 default_stages: [push]
 
 repos:
+  - repo: https://github.com/codespell-project/codespell
+    rev: v2.2.2
+    hooks:
+      - id: codespell
+        args: [
+            "-w", # Write changes to files
+          ]
+
   - repo: https://github.com/pycqa/isort
     rev: 5.12.0
     hooks:
@@ -38,9 +46,4 @@ repos:
             "-sn", # Don't display the score
             "--disable=R,import-error", # Refactors are not important enough to block a commit.
           ]
-
-  - repo: https://github.com/codespell-project/codespell
-    rev: v2.2.2
-    hooks:
-      - id: codespell
 # Unused-imports aren't testable by the github action, so don't test that here.
