+++ pyproject.toml
@@ -38,7 +38,7 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
     "pytest>=7.1.3, <7.1.5",
-    "black>=22.8.0,<22.12.1",
+    "black>=22.8.0,<23.1.1",
     "pre-commit>=2.20.0,<3.1.1",
     "pytest-cov>=3.0.0,<4.0.1",
     "flake8>=5.0.0, <5.0.6",
