### 2023-03-06 8a8eaad
fix:(deps-dev): update srsly requirement
Updates the requirements on [srsly](https://github.com/explosion/srsly) to permit the latest version.
- [Release notes](https://github.com/explosion/srsly/releases)
- [Commits](https://github.com/explosion/srsly/compare/v2.4.4...v2.4.6)

---
updated-dependencies:
- dependency-name: srsly
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
