+++ pyproject.toml
@@ -17,7 +17,7 @@ dependencies = [
     "wandb>=0.12.0,<0.13.8",
     "deepchecks>=0.8.0,<0.12.0",
     "dill>=0.3.0,<0.3.7",
-    "srsly>=2.4.4,<2.4.6",
+    "srsly>=2.4.4,<2.4.7",
     "SQLAlchemy>=1.4.41, <1.5.42",
     "pyodbc>=4.0.34, <4.0.36",
     "dask>=2022.9.0,<2023.2.0",
