### 2023-03-06 4308f96
fix:(deps-dev): update pylint requirement
Updates the requirements on [pylint](https://github.com/PyCQA/pylint) to permit the latest version.
- [Release notes](https://github.com/PyCQA/pylint/releases)
- [Commits](https://github.com/PyCQA/pylint/compare/v2.15.5...v2.16.3)

---
updated-dependencies:
- dependency-name: pylint
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
