+++ pyproject.toml
@@ -45,7 +45,7 @@ dev = [
     "docformatter>=1.5.0, <1.5.2",
     "mypy>=0.971,<0.992",
     "pytest-xdist>=2.4.0,<3.1.1",
-    "pylint>=2.15.5,<2.16.0",
+    "pylint>=2.15.5,<2.17.0",
 ]
 
 docs = [
