+++ pyproject.toml
@@ -37,7 +37,7 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = [
-    "pytest>=7.1.3, <7.1.5",
+    "pytest>=7.1.3, <7.2.3",
     "black>=22.8.0,<23.1.1",
     "pre-commit>=2.20.0,<3.1.1",
     "pytest-cov>=3.0.0,<4.0.1",
