### 2023-03-07 7478edc
Make 3.9 to be minimal supported version as manuscript says
may be that is not what you want really since 3.8 is still widely used
and supported by Python. But only 3.9 is tested ATM
