+++ pyproject.toml
@@ -8,7 +8,7 @@ authors = [{name = "Martin Bernstorff", email = "martinbernstorff@gmail.com"},
     {name = "Jakob Grøhn Damgaard", email = "bokajgd@gmail.com"},
     {name = "Kenneth Enevoldsen"}]
 
-requires-python = ">=3.8.0"
+requires-python = ">=3.9.0"
 
 dependencies = [
     "scipy>=1.8.0,<1.9.4",
