### 2023-03-10 1004142
feat: check if entity ids unique for incident=true
