### 2023-03-10 ab4fc2b
ci: add python 3.9, 3.10 and 3.11 to tests
