+++ tests/test_timeseriesflattener/test_feature_spec_objects.py
@@ -236,4 +236,6 @@ def test_outcome_spec_incident_true(caplog):
         feature_name="value",
     )
 
-    assert "Incident outcomes must have unique IDs. You have set incident=[True], however, the keys in your ID column are not unique. Consider setting incident=[False]." in caplog.text
+    assert (
+        "Incident outcomes must have unique IDs. You have set incident=[True], however, the keys in your ID column are not unique. Consider setting incident=[False]." in caplog.text
+        )
