### 2023-03-10 49b58e4
ci: run jupyter tutorials on correct python version
