+++ .github/workflows/main_test_and_release.yml
@@ -42,7 +42,7 @@ jobs:
   jupyter-tutorials:
     runs-on: ${{ matrix.os }}
     env:
-      python-version: ${{ matrix.os }}
+      python-version: ${{ matrix.python-version }}
 
     strategy:
       matrix:
