+++ .github/workflows/main_test_and_release.yml
@@ -20,14 +20,14 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        python-version: [3.9, 3.10, 3.11]
+        python-version: ["3.9", "3.10", "3.11"]
 
     env:
       python-version: ${{ matrix.python-version }}
 
     # This allows a subsequently queued workflow run to interrupt previous runs
     concurrency:
-      group: "${{ github.workflow }} - ${{ matrix.os }} @ ${{ github.ref }}"
+      group: "${{ github.workflow }} - ${{ matrix.python-version}} ${{ matrix.os }} @ ${{ github.ref }}"
       cancel-in-progress: true
 
     steps:
