### 2023-03-10 e24a06b
ci: attempt support for python 3.8
