### 2023-03-10 5042886
tests: disable remaining huggingface test
