+++ tests/test_timeseriesflattener/test_flattened_dataset/test_flattened_dataset.py
@@ -2,7 +2,6 @@
 import numpy as np
 import pandas as pd
 import pytest
-
 from timeseriesflattener.feature_spec_objects import (
     OutcomeSpec,
     PredictorSpec,
@@ -14,7 +13,6 @@ from timeseriesflattener.resolve_multiple_functions import latest, mean
 from timeseriesflattener.testing.utils_for_testing import (
     synth_outcome,
     synth_prediction_times,
-    synth_text_data,
 )
 from timeseriesflattener.text_embedding_functions import sentence_transformers_embedding
 
@@ -66,7 +64,7 @@ def test_add_spec(synth_prediction_times: pd.DataFrame, synth_outcome: pd.DataFr
     with pytest.raises(ValueError):
         dataset.add_spec("invalid spec")
 
-
+@pytest.mark.huggingface
 def test_compute_specs(
     synth_prediction_times: pd.DataFrame,
     synth_outcome: pd.DataFrame,
