+++ pyproject.toml
@@ -37,13 +37,13 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = [
-    "pytest>=7.1.3, <7.2.3",
     "black>=22.8.0,<23.1.1",
     "pre-commit>=2.20.0,<3.1.1",
-    "pytest-cov>=3.0.0,<4.0.1",
     "flake8>=5.0.0, <5.0.6",
     "docformatter>=1.5.0, <1.5.2",
     "mypy>=0.971,<0.992",
+    "pytest>=7.1.3, <7.2.3",
+    "pytest-cov>=3.0.0,<4.0.1",
     "pytest-xdist>=2.4.0,<3.1.1",
     "pylint>=2.15.5,<2.17.0",
 ]
