### 2023-03-10 50c5aca
fix: sklearn 1.2 compatibility
