### 2023-03-10 67e3e32
build: remove 3.8, we require functools cache
