### 2023-03-10 c628c61
refactor: changed warning to ValueError
