### 2023-03-10 655746c
tests: simplify tests and add test for lack of error
