+++ pyproject.toml
@@ -13,7 +13,7 @@ requires-python = ">=3.9.0,<3.11.0"
 dependencies = [
     "scipy>=1.8.0,<1.9.4",
     "wasabi>=0.9.1,<1.1.2",
-    "scikit-learn>=1.1.2,<1.2.2",
+    "scikit-learn>=1.1.2,<1.2.3",
     "wandb>=0.12.0,<0.13.11",
     "deepchecks>=0.8.0,<0.12.0",
     "dill>=0.3.0,<0.3.7",
