### 2023-03-20 d874eb9
fix:(deps-dev): update wandb requirement
Updates the requirements on [wandb](https://github.com/wandb/wandb) to permit the latest version.
- [Release notes](https://github.com/wandb/wandb/releases)
- [Changelog](https://github.com/wandb/wandb/blob/main/CHANGELOG.md)
- [Commits](https://github.com/wandb/wandb/compare/v0.12.0...v0.14.0)

---
updated-dependencies:
- dependency-name: wandb
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
