+++ pyproject.toml
@@ -14,7 +14,7 @@ dependencies = [
     "scipy>=1.8.0,<1.9.4",
     "wasabi>=0.9.1,<1.1.2",
     "scikit-learn>=1.1.2,<1.2.3",
-    "wandb>=0.12.0,<0.13.12",
+    "wandb>=0.12.0,<0.14.1",
     "deepchecks>=0.8.0,<0.14.0",
     "dill>=0.3.0,<0.3.7",
     "srsly>=2.4.4,<2.4.7",
