+++ pyproject.toml
@@ -44,7 +44,7 @@ dev = [
     "mypy>=0.971,<0.992",
     "pytest>=7.1.3, <7.2.3",
     "pytest-cov>=3.0.0,<4.0.1",
-    "pytest-xdist>=2.4.0,<3.1.1",
+    "pytest-xdist>=2.4.0,<3.2.2",
     "pylint>=2.15.5,<2.17.0",
 ]
 
