+++ pyproject.toml
@@ -21,7 +21,7 @@ dependencies = [
     "SQLAlchemy>=1.4.41, <1.5.42",
     "pyodbc>=4.0.34, <4.0.36",
     "dask>=2022.9.0,<2023.4.0",
-    "pydantic>=1.9.0, <1.10.0",
+    "pydantic>=1.9.0, <1.11.0",
     "psutil>=5.9.1, <6.0.0",
     "pandas>=1.4.0,<1.6.0",
     "catalogue>=2.0.0, <2.1.0",
