### 2023-03-20 060b7fd
ci: minimise pre-commit
