+++ .pre-commit-config.yaml
@@ -1,49 +1,27 @@
 default_stages: [push]
 
 repos:
-  - repo: https://github.com/codespell-project/codespell
-    rev: v2.2.2
-    hooks:
-      - id: codespell
-        args: [
-            "-w", # Write changes to files
-          ]
-
-  - repo: https://github.com/pycqa/isort
-    rev: 5.12.0
-    hooks:
-      - id: isort
-        name: isort (python)
-        args: ["--profile", "black", "--filter-files"]
-
-  - repo: https://github.com/asottile/pyupgrade
-    rev: v2.38.0
-    hooks:
-      - id: pyupgrade
-
-  - repo: https://github.com/bwhmather/ssort
-    rev: v0.11.6
-    hooks:
-      - id: ssort
-
-  - repo: https://github.com/asottile/add-trailing-comma
-    rev: v2.2.3
-    hooks:
-      - id: add-trailing-comma
-
   - repo: https://github.com/psf/black
     rev: 22.8.0
     hooks:
       - id: black
 
-  - repo: https://github.com/pycqa/pylint
-    rev: v2.15.2
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    # Ruff version.
+    rev: v0.0.254
     hooks:
-      - id: pylint
-        types: [python]
-        args: [
-            "-rn", # Only display messages
-            "-sn", # Don't display the score
-            "--disable=R,import-error", # Refactors are not important enough to block a commit.
+      - id: ruff
+        args:
+          [
+            "--extend-select",
+            "F401",
+            "--extend-select",
+            "F841",
+            "--fix",
+            "--exit-non-zero-on-fix",
           ]
-# Unused-imports aren't testable by the github action, so don't test that here.
+
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v2.3.0
+    hooks:
+      - id: check-yaml
