### 2023-03-20 fd80f5a
misc: prepare for python 3.8 support
