+++ tests/test_timeseriesflattener/test_flattened_dataset/test_concatenation.py
@@ -2,11 +2,10 @@
 import random
 import time
 import uuid
-from typing import Callable
+from typing import Callable, List
 
 import pandas as pd
 import pytest
-
 from timeseriesflattener.flattened_dataset import TimeseriesFlattener
 
 # pylint: disable=protected-access
@@ -21,7 +20,7 @@ def benchmark(func: Callable, *args, **kwargs):
     return end - start
 
 
-def generate_test_df(uuids: list[str], col_values: list[int] = None):
+def generate_test_df(uuids: List[str], col_values: List[int] = None):
     """Generate a test df with a random integer column and a uuid index."""
     df = pd.DataFrame()
 
