+++ src/timeseriesflattener/flattened_dataset.py
@@ -6,7 +6,7 @@ import datetime as dt
 import logging
 import random
 import time
-from collections.abc import Callable
+typing import Callable
 from datetime import timedelta
 from multiprocessing import Pool
 from typing import List, Optional, Union
@@ -15,7 +15,7 @@ import coloredlogs
 import numpy as np
 import pandas as pd
 import tqdm
-from catalogue import Registry  # noqa # pylint: disable=unused-import
+from catalogue import Registry
 from dask.diagnostics import ProgressBar
 from pandas import DataFrame
 from pydantic import BaseModel as PydanticBaseModel
@@ -42,9 +42,9 @@ log = logging.getLogger(__name__)
 class SpecCollection(PydanticBaseModel):
     """A collection of specs."""
 
-    outcome_specs: list[OutcomeSpec] = []
-    predictor_specs: list[PredictorSpec] = []
-    static_specs: list[_AnySpec] = []
+    outcome_specs: List[OutcomeSpec] = []
+    predictor_specs: List[PredictorSpec] = []
+    static_specs: List[_AnySpec] = []
 
     def __len__(self):
         """Return number of specs in collection."""
@@ -93,14 +93,14 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
     def __init__(  # pylint: disable=too-many-arguments
         self,
         prediction_times_df: DataFrame,
-        drop_pred_times_with_insufficient_look_distance: bool,  # noqa
+        drop_pred_times_with_insufficient_look_distance: bool,
         cache: Optional[FeatureCache] = None,
         entity_id_col_name: str = "entity_id",
         timestamp_col_name: str = "timestamp",
         predictor_col_name_prefix: str = "pred",
         outcome_col_name_prefix: str = "outc",
         n_workers: int = 60,
-        log_to_stdout: bool = True,  # noqa
+        log_to_stdout: bool = True,
     ):
         """Class containing a time-series, flattened.
 
@@ -245,9 +245,6 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
         # Sort by timestamp_pred in case resolve_multiple needs dates
         df = df.sort_values(by=val_timestamp_col_name).groupby(pred_time_uuid_colname)
 
-        if isinstance(resolve_multiple, str):
-            resolve_multiple = resolve_multiple_fns.get(resolve_multiple)
-
         if callable(resolve_multiple):
             df = resolve_multiple(df).reset_index()
         else:
@@ -304,13 +301,13 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
         )
 
     @staticmethod
-    def _flatten_temporal_values_to_df(  # noqa pylint: disable=too-many-locals
+    def _flatten_temporal_values_to_df(
         prediction_times_with_uuid_df: DataFrame,
         output_spec: TemporalSpec,
         entity_id_col_name: str,
         pred_time_uuid_col_name: str,
         timestamp_col_name: str,
-        verbose: bool = False,  # noqa
+        verbose: bool = False,
     ) -> DataFrame:
         """Create a dataframe with flattened values (either predictor or
 
@@ -417,7 +414,7 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
             pred_time_uuid_colname=pred_time_uuid_col_name,
         ).fillna(output_spec.fallback)
 
-        return df[value_col_str_name + [pred_time_uuid_col_name]]
+        return df[[*value_col_str_name, pred_time_uuid_col_name]]
 
     @staticmethod
     def rename_input_col_to_value(
@@ -475,16 +472,10 @@ class TimeseriesFlattener:  # pylint: disable=too-many-instance-attributes
                 df=df,
             )
 
[…]
