+++ src/timeseriesflattener/text_embedding_functions.py
@@ -1,5 +1,5 @@
 """Functions for embedding text data"""
-from collections.abc import Callable
+typing import Callable
 from typing import Optional
 
 import pandas as pd

[…]
