+++ tests/test_timeseriesflattener/test_feature_spec_objects.py
@@ -1,10 +1,8 @@
 """Test that feature spec objects work as intended."""
-import difflib
 
 import numpy as np
 import pandas as pd
 import pytest
-
 from timeseriesflattener.feature_spec_objects import (
     BaseModel,
     OutcomeGroupSpec,
@@ -22,8 +20,7 @@ from timeseriesflattener.testing.load_synth_data import (  # pylint: disable=unu
     load_synth_predictor_float,
     synth_predictor_binary,
 )
-from timeseriesflattener.testing.utils_for_testing import long_df_with_multiple_values
-from timeseriesflattener.utils import data_loaders, split_df_and_register_to_dict
+from timeseriesflattener.utils import split_df_and_register_to_dict
 
 
 def test_anyspec_init():
