+++ tests/test_timeseriesflattener/test_embedding_functions.py
@@ -1,8 +1,6 @@
 """Tests for the text embedding functions"""
 import pytest
-
 from timeseriesflattener.testing.text_embedding_functions import bow_test_embedding
-from timeseriesflattener.testing.utils_for_testing import synth_text_data
 from timeseriesflattener.text_embedding_functions import (
     huggingface_embedding,
     sentence_transformers_embedding,
@@ -16,7 +14,7 @@ def test_embedding_fn(synth_text_data):
     assert embedding_df.shape == (df.shape[0], 10)
 
 
-@pytest.mark.huggingface
+@pytest.mark.huggingface()
 def test_huggingface_embedding(synth_text_data):
     """Test that the huggingface embedding function works as expected"""
     df = synth_text_data.dropna(subset="text")
@@ -28,7 +26,7 @@ def test_huggingface_embedding(synth_text_data):
     assert embedding_df.shape == (df.shape[0], 384)
 
 
-@pytest.mark.huggingface
+@pytest.mark.huggingface()
 def test_sentence_transformer_embedding(synth_text_data):
     """Test that the sentence-transformer embedding function works as expected"""
     df = synth_text_data.dropna(subset="text")
