+++ src/timeseriesflattener/feature_spec_objects.py
@@ -2,9 +2,8 @@
 import itertools
 import logging
 import time
-typing import Callable, Sequence
 from functools import lru_cache
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Callable, Dict, List, Optional, Sequence, Union
 
 import pandas as pd
 from frozendict import frozendict
@@ -114,7 +113,7 @@ class BaseModel(PydanticBaseModel):
     # This means we want to auto-generate docstrings to support the inheritance,
     # but also need to hard-code the docstring to support pylance.
     class Doc:
-        short_description: str = """Modified Pydantic BaseModel to allow arbitrary 
+        short_description: str = """Modified Pydantic BaseModel to allow arbitrary
         types and disallow attributes not in the class."""
 
     class Config:
@@ -224,8 +223,8 @@ class _AnySpec(BaseModel):
 
     values_name: Optional[str] = Field(
         default=None,
-        description="""A string that maps to a key in a dictionary instantiated by 
-            `split_df_and_register_to_dict`. Each key corresponds to a dataframe, which 
+        description="""A string that maps to a key in a dictionary instantiated by
+            `split_df_and_register_to_dict`. Each key corresponds to a dataframe, which
             is a subset of the df where the values_name == key.""",
     )
 
@@ -240,7 +239,7 @@ class _AnySpec(BaseModel):
     )
 
     feature_name: str = Field(
-        description="""The name of the feature. Used for column name generation, e.g. 
+        description="""The name of the feature. Used for column name generation, e.g.
             <prefix>_<feature_name>.""",
     )
 
@@ -251,8 +250,8 @@ class _AnySpec(BaseModel):
 
     input_col_name_override: Optional[str] = Field(
         default=None,
-        description="""An override for the input column name. If None, will  attempt 
-            to infer it by looking for the only column that doesn't match id_col_name 
+        description="""An override for the input column name. If None, will  attempt
+            to infer it by looking for the only column that doesn't match id_col_name
             or timestamp_col_name.""",
     )
 
@@ -373,23 +372,23 @@ class TemporalSpec(_AnySpec):
     """
 
     class Doc:
-        short_description = """The minimum specification required for collapsing a temporal 
+        short_description = """The minimum specification required for collapsing a temporal
         feature, whether looking ahead or behind. Mostly used for inheritance below."""
 
     interval_days: Union[int, float] = Field(
-        description="""How far to look in the given direction (ahead for outcomes, 
+        description="""How far to look in the given direction (ahead for outcomes,
             behind for predictors)""",
     )
 
     resolve_multiple_fn: Union[Callable, str] = Field(
-        description="""A function used for resolving multiple values within the 
+        description="""A function used for resolving multiple values within the
             interval_days.""",
     )
 
     key_for_resolve_multiple: Optional[str] = Field(
         default=None,
-        description="""Key used to lookup the resolve_multiple_fn in the 
-            resolve_multiple_fns registry. Used for column name generation. Only 
+        description="""Key used to lookup the resolve_multiple_fn in the
+            resolve_multiple_fns registry. Used for column name generation. Only
             required if you don't specify a resolve_multiple_fn. Call
             timeseriesflattener.resolve_multiple_fns.resolve_multiple_fns.get_all()
             for a list of options.""",
@@ -401,7 +400,7 @@ class TemporalSpec(_AnySpec):
 
     allowed_nan_value_prop: float = Field(
         default=0.0,
-        description="""If NaN is higher than this in the input dataframe during 
+        description="""If NaN is higher than this in the input dataframe during
             resolution, raise an error.""",
     )
 
@@ -591,7 +590,7 @@ class TextPredictorSpec(PredictorSpec):
         )
 
     embedding_fn: Callable = Field(
[…]
