+++ src/timeseriesflattener/testing/text_embedding_functions.py
@@ -6,7 +6,6 @@ import pandas as pd
 from pandas import DataFrame, Series
 from sklearn.decomposition import PCA
 from sklearn.feature_extraction.text import CountVectorizer
-
 from timeseriesflattener.utils import PROJECT_ROOT
 
 
