### 2023-03-20 dfa91ad
ci: ignore all prefixed with .venv and check pyproject
