+++ .pre-commit-config.yaml
@@ -25,4 +25,8 @@ repos:
     rev: v2.3.0
     hooks:
       - id: check-yaml
-      - id: check-toml
+
+  - repo: https://github.com/repo-helper/pyproject-parser
+    rev: v0.7.0
+    hooks:
+      - id: check-pyproject
