### 2023-03-20 2a46071
tests: move the rest of the fixtures to conftest.py
