+++ tests/test_timeseriesflattener/test_utils.py
@@ -1,7 +1,6 @@
 """Test that feature spec objects work as intended."""
 import pandas as pd
 from timeseriesflattener.feature_spec_objects import _AnySpec
-from timeseriesflattener.testing.load_synth_data import synth_predictor_binary  # noqa
 from timeseriesflattener.utils import split_df_and_register_to_dict, split_dfs
 
 
