+++ tests/test_timeseriesflattener/test_flattened_dataset/test_flattened_dataset.py
@@ -10,16 +10,7 @@ from timeseriesflattener.feature_spec_objects import (
 )
 from timeseriesflattener.flattened_dataset import TimeseriesFlattener
 from timeseriesflattener.resolve_multiple_functions import latest, mean
-from timeseriesflattener.testing.utils_for_testing import (
-    synth_outcome,
-    synth_prediction_times,
-)
 from timeseriesflattener.text_embedding_functions import sentence_transformers_embedding
-
-# To avoid ruff auto-removing unused imports
-used_funcs = [synth_prediction_times, synth_outcome]
-
-# pylint: disable=missing-function-docstring
 
 
 def test_add_spec(synth_prediction_times: pd.DataFrame, synth_outcome: pd.DataFrame):
