+++ .github/workflows/main_test_and_release.yml
@@ -12,7 +12,7 @@ on:
       - main
 
 env:
-  cache-version: 0.0.8 # Change this number if you want to manually invalidate all caches
+  cache-version: 0.0.9 # Change this number if you want to manually invalidate all caches
 
 jobs:
   test:
