+++ tests/conftest.py
@@ -1,15 +1,13 @@
-# pylint: disable-all
-
-from typing import Optional
-
-import pandas as pd
 import pytest
 from pandas import DataFrame
-from timeseriesflattener.testing.load_synth_data import load_raw_test_csv
+from timeseriesflattener.testing.load_synth_data import (
+    load_synth_outcome,
+    load_synth_prediction_times,
+    load_synth_text,
+)
 from timeseriesflattener.testing.utils_for_testing import (
     load_long_df_with_multiple_values,
 )
-from timeseriesflattener.utils import data_loaders
 
 
 def pytest_addoption(parser):  # noqa
@@ -66,99 +64,3 @@ def long_df_with_multiple_values() -> DataFrame:
 def synth_text_data() -> DataFrame:
     """Load the synth text data."""
     return load_synth_text()
-
-
-@data_loaders.register("synth_predictor_float")
-def load_synth_predictor_float(
-    n_rows: Optional[int] = None,
-) -> pd.DataFrame:
-    """Load synth predictor data.".
-
-    Args:
-        n_rows: Number of rows to return. Defaults to None which returns entire coercion data view.
-
-    Returns:
-        pd.DataFrame
-    """
-    return load_raw_test_csv("synth_raw_float_1.csv", n_rows=n_rows)
-
-
-@data_loaders.register("synth_sex")
-def load_synth_sex(
-    n_rows: Optional[int] = None,
-) -> pd.DataFrame:
-    """Load synth sex data.".
-
-    Args:
-        n_rows: Number of rows to return. Defaults to None which returns entire coercion data view.
-
-    Returns:
-        pd.DataFrame
-    """
-    return load_raw_test_csv("synth_sex.csv", n_rows=n_rows)
-
-
-@data_loaders.register("synth_predictor_binary")
-def synth_predictor_binary(
-    n_rows: Optional[int] = None,
-) -> pd.DataFrame:
-    """Load synth predictor data.".
-
-    Args:
-        n_rows: Number of rows to return. Defaults to None which returns entire coercion data view.
-
-    Returns:
-        pd.DataFrame
-    """
-    return load_raw_test_csv("synth_raw_binary_1.csv", n_rows=n_rows)
-
-
-@data_loaders.register("synth_outcome")
-def load_synth_outcome(
-    n_rows: Optional[int] = None,
-) -> pd.DataFrame:
-    """Load synth predictor data.".
-
-    Args:
-        n_rows: Number of rows to return. Defaults to None which returns entire coercion data view.
-
-    Returns:
-        pd.DataFrame
-    """
-    # Get first row for each id
-    df = load_raw_test_csv("synth_raw_binary_2.csv", n_rows=n_rows)
-    df = df.groupby("entity_id").last().reset_index()
-
-    # Drop all rows with a value equal to 1
-    df = df[df["value"] == 1]
-    return df
-
-
-@data_loaders.register("synth_prediction_times")
-def load_synth_prediction_times(
-    n_rows: Optional[int] = None,
-) -> pd.DataFrame:
-    """Load synth predictor data.".
-
-    Args:
-        n_rows: Number of rows to return. Defaults to None which returns entire coercion data view.
-
-    Returns:
-        pd.DataFrame
-    """
-    return load_raw_test_csv("synth_prediction_times.csv", n_rows=n_rows)
-
-
-@data_loaders.register("synth_text")
-def load_synth_text(
-    n_rows: Optional[int] = None,
-) -> pd.DataFrame:
-    """Load synth text data.".
-
-    Args:
-        n_rows: Number of rows to return. Defaults to None which returns entire coercion data view.
-
-    Returns:
-        pd.DataFrame
-    """
-    return load_raw_test_csv("synth_text_data.csv", n_rows=n_rows)
