### 2023-03-20 9ed0800
docs: add link to pypi
