+++ tests/test_data/raw/create_synth_raw_binary.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from psycopmlutils.synth_data_generator.synth_col_generators import (
+from psycop_ml_utilssynth_data_generator.synth_col_generators import (
     generate_data_columns,
 )
 
