+++ tests/test_data/flattened/generated_with_outcome/create_synth_flattened_with_outcome.py
@@ -2,7 +2,7 @@
 from pathlib import Path
 
 import numpy as np
-from psycopmlutils.synth_data_generator.synth_prediction_times_generator import (
+from psycop_ml_utilssynth_data_generator.synth_prediction_times_generator import (
     generate_synth_data,
 )
 
