+++ tests/test_data/raw/create_synth_prediction_times.py
@@ -1,6 +1,6 @@
 """Generate dataframe with prediction times."""
 
-from psycop_ml_utilssynth_data_generator.synth_col_generators import (
+from psycop_ml_utils.synth_data_generator.synth_col_generators import (
     generate_data_columns,
 )
 from timeseriesflattener.utils import PROJECT_ROOT
