+++ tests/test_data/raw/create_synth_raw_float.py
@@ -1,6 +1,6 @@
 """Generate raw float dataframe."""
 
-from psycop_ml_utilssynth_data_generator.synth_col_generators import (
+from psycop_ml_utils.synth_data_generator.synth_col_generators import (
     generate_data_columns,
 )
 from timeseriesflattener.utils import PROJECT_ROOT
