### 2023-03-20 b465d2f
fix: allow newest 3.10 version
