+++ pyproject.toml
@@ -16,7 +16,7 @@ classifiers = [
   "Programming Language :: Python :: 3.10"
 ]
 
-requires-python = ">=3.8.0,<=3.10.0"
+requires-python = ">=3.8.0,<=3.11.0"
 
 dependencies = [
     "scipy>=1.8.0,<1.9.4",
