+++ pyproject.toml
@@ -3,10 +3,7 @@ name = "timeseriesflattener"
 version = "0.23.4"
 description = "A package for converting time series data from e.g. electronic health records into wide format data."
 readme = "README.md"
-authors = [{name = "Martin Bernstorff", email = "martinbernstorff@gmail.com"},
-    {name = "Lasse Hansen", email = "lasseh0310@gmail.com"},
-    {name = "Jakob Grøhn Damgaard", email = "bokajgd@gmail.com"},
-    {name = "Kenneth Enevoldsen"}]
+authors = [{name = "Lasse Hansen", email = "lasseh0310@gmail.com"}, {name = "Jakob Grøhn Damgaard", email = "bokajgd@gmail.com"}, {name = "Kenneth Enevoldsen"}, {name = "Martin Bernstorff", email = "martinbernstorff@gmail.com"}]
 classifiers = [
   "Operating System :: POSIX :: Linux",
   "Operating System :: MacOS :: MacOS X",
@@ -16,7 +13,7 @@ classifiers = [
   "Programming Language :: Python :: 3.10"
 ]
 
-requires-python = ">=3.8.0,<=3.11.0"
+requires-python = ">=3.8.0,<3.11.0"
 
 dependencies = [
     "scipy>=1.8.0,<1.9.4",
