### 2023-03-20 ca48492
build: narrow down and remove unused dependencies
