+++ .github/actions/test_tutorials/action.yml
@@ -4,7 +4,7 @@ inputs:
   python-version:
     description: "Which Python version to run on"
     required: true
-    default: 3.9
+    default: "3.9"
 runs:
   using: "composite"
   steps:
@@ -19,7 +19,7 @@ runs:
       shell: bash
       run: |
         python -m pip install --upgrade pip
-        pip install -e ".[dev,text]"
+        pip install -e ".[dev,text,tutorials]"
 
     ### Run tests
     - name: Install jupyter
