+++ pyproject.toml
@@ -14,7 +14,7 @@ classifiers = [
   "Programming Language :: Python :: 3.11"
 ]
 
-requires-python = ">=3.8.0,<3.11.0"
+requires-python = ">=3.8.0,<3.12.0"
 
 dependencies = [
     "scipy>=1.8.0,<1.9.4",
