### 2023-03-20 #5197 @MartinBernstorff
> @MartinBernstorff confusion with python 3.11 support remains. Filed issue(s).

Thanks a lot for the nudge, @yarikoptic! Did some work on dependencies, support now covers 3.8, 3.9, 3.10 and 3.11.
