### 2023-03-20 6f7959b
fix: update supported python versions on pypi
