### 2023-03-22 Aarhus-Psychiatry-Research/timeseriesflattener#154 @mstimberg
Paper: Please add captions to the tables
[Finalizing the review for JOSS: https://github.com/openjournals/joss-reviews/issues/5197]

Could you please add (short) captions to the two tables in the paper so that they get numbered as Table 1 and Table 2 (since you refer to them earlier):
> A caption is a paragraph beginning with the string Table: (or table: or just :), which will be stripped off. It may appear either before or after the table.

https://pandoc.org/MANUAL.html#extension-table_captions
