### 2023-03-22 Aarhus-Psychiatry-Research/timeseriesflattener#155 @mstimberg
Paper: minor reference fixes
[Finalizing the review for JOSS: https://github.com/openjournals/joss-reviews/issues/5197]

- Could you please add a "References" section header before the reference list
- The reference for XGBboost looks a bit odd and references a rather old version. Since you seem to cite it in a general way, I think it would be better to use the citation that the developers suggest: https://github.com/dmlc/xgboost/blob/master/CITATION
- Could you add the DOI (`10.1109/RBME.2020.3007816`) to the reference for the Shamout et al. IEEE paper? The correct year also seems to be 2021not 2020 (https://ieeexplore.ieee.org/document/9134853)
