### 2023-03-22 Aarhus-Psychiatry-Research/timeseriesflattener#156 @mstimberg
Paper: updated supported Python version
[Finalizing the review for JOSS: https://github.com/openjournals/joss-reviews/issues/5197]

The paper still states "timeseriesflattener is a Python package (3.8 | 3.9 | 3.10)", but if I understood correctly, it is now
compatible with Python 3.11 as well.

Really minor/optional thing: The "tutorials" link currently points to the github repository, maybe pointing it to https://aarhus-psychiatry-research.github.io/timeseriesflattener/tutorials.html would be more appropriate?
