### 2023-03-23 Aarhus-Psychiatry-Research/timeseriesflattener#154 @MartinBernstorff
@HLasse Will you take a look? :-)
