+++ paper/paper.bib
@@ -1,140 +1,146 @@
 
 @article{rajkomar_scalable_2018,
-	title = {Scalable and accurate deep learning with electronic health records},
-	volume = {1},
-	rights = {2018 The Author(s)},
-	issn = {2398-6352},
-	url = {https://www.nature.com/articles/s41746-018-0029-1/},
-	doi = {10.1038/s41746-018-0029-1},
-	abstract = {Predictive modeling with electronic health record ({EHR}) data is anticipated to drive personalized medicine and improve healthcare quality. Constructing predictive statistical models typically requires extraction of curated predictor variables from normalized {EHR} data, a labor-intensive process that discards the vast majority of information in each patient’s record. We propose a representation of patients’ entire raw {EHR} records based on the Fast Healthcare Interoperability Resources ({FHIR}) format. We demonstrate that deep learning methods using this representation are capable of accurately predicting multiple medical events from multiple centers without site-specific data harmonization. We validated our approach using de-identified {EHR} data from two {US} academic medical centers with 216,221 adult patients hospitalized for at least 24 h. In the sequential format we propose, this volume of {EHR} data unrolled into a total of 46,864,534,945 data points, including clinical notes. Deep learning models achieved high accuracy for tasks such as predicting: in-hospital mortality (area under the receiver operator curve [{AUROC}] across sites 0.93–0.94), 30-day unplanned readmission ({AUROC} 0.75–0.76), prolonged length of stay ({AUROC} 0.85–0.86), and all of a patient’s final discharge diagnoses (frequency-weighted {AUROC} 0.90). These models outperformed traditional, clinically-used predictive models in all cases. We believe that this approach can be used to create accurate and scalable predictions for a variety of clinical scenarios. In a case study of a particular prediction, we demonstrate that neural networks can be used to identify relevant information from the patient’s chart.},
-	pages = {1--10},
-	number = {1},
-	journaltitle = {npj Digital Medicine},
-	author = {Rajkomar, Alvin and Oren, Eyal and Chen, Kai and Dai, Andrew M. and Hajaj, Nissan and Hardt, Michaela and Liu, Peter J. and Liu, Xiaobing and Marcus, Jake and Sun, Mimi and Sundberg, Patrik and Yee, Hector and Zhang, Kun and Zhang, Yi and Flores, Gerardo and Duggan, Gavin E. and Irvine, Jamie and Le, Quoc and Litsch, Kurt and Mossin, Alexander and Tansuwan, Justin and Wang, De and Wexler, James and Wilson, Jimbo and Ludwig, Dana and Volchenboum, Samuel L. and Chou, Katherine and Pearson, Michael and Madabushi, Srinivasan and Shah, Nigam H. and Butte, Atul J. and Howell, Michael D. and Cui, Claire and Corrado, Greg S. and Dean, Jeffrey},
-	urldate = {2020-09-13},
-	date = {2018-05-08},
-	langid = {english},
-	note = {Number: 1
-Publisher: Nature Publishing Group},
-	file = {Full Text PDF:/Users/au554730/Zotero/storage/XTCIFGTX/Rajkomar et al. - 2018 - Scalable and accurate deep learning with electroni.pdf:application/pdf;Snapshot:/Users/au554730/Zotero/storage/DC6GJ9PM/s41746-018-0029-1.html:text/html},
+  title        = {Scalable and accurate deep learning with electronic health records},
+  volume       = {1},
+  rights       = {2018 The Author(s)},
+  issn         = {2398-6352},
+  url          = {https://www.nature.com/articles/s41746-018-0029-1/},
+  doi          = {10.1038/s41746-018-0029-1},
[…]
