### 2023-03-23 Aarhus-Psychiatry-Research/timeseriesflattener#156 @MartinBernstorff
Thanks a ton for the attention to detail! Fixing now.
