### 2023-03-23 Aarhus-Psychiatry-Research/timeseriesflattener#155 @MartinBernstorff
Excellent points, fixing now.
