+++ paper/paper.bib
@@ -114,6 +114,7 @@
   pages        = {116--126},
   journaltitle = {{IEEE} reviews in Biomedical Engineering},
   author       = {Shamout, Farah and Zhu, Tingting and Clifton, David A.},
+  doi          = {10.1109/RBME.2020.3007816},
   date         = {2020},
   note         = {Publisher: {IEEE}}
 }
