### 2023-03-23 1e8b163
ci: run paper action anytime the paper markdown file has changed
