+++ .github/workflows/generate_paper_pdf.yml
@@ -2,8 +2,8 @@ name: Create paper pdf
 
 on:
   push:
-    branches:
-      - main 
+    paths:
+      - "paper/*.md"
 
 jobs:
   paper:
