### 2023-03-23 Aarhus-Psychiatry-Research/timeseriesflattener#155 @mstimberg
Looks good in general, but the Shamout et al. reference still has the year as 2020 instead of 2021 in the `.bib` file, and its citation is currently broken (should use `@shamout_machine_2021` instead of `@shamout_machine_2020` now).
