### 2023-03-24 #5197 @MartinBernstorff
> 👋 @MartinBernstorff As you've probably seen, the two reviewers have finished their reviews and recommend accepting the software/paper, congratulations 🎉
>
> I've filed three minor issues regarding the paper ([Aarhus-Psychiatry-Research/timeseriesflattener#154](https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener/issues/154), [Aarhus-Psychiatry-Research/timeseriesflattener#155](https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener/issues/155), [Aarhus-Psychiatry-Research/timeseriesflattener#156](https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener/issues/156)). As soon as they are fixed (assuming no errors/misunderstandings on my side, of course), we can proceed with finalizing the acceptance and publishing process.

Thansk @mstimberg! I've fixed the issues, thanks for the attention to detail 👍
