### 2023-03-24 #5197 @mstimberg
@MartinBernstorff great, thanks for the changes. There's still a minor issue with one reference, see my comment at https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener/issues/155.
