+++ paper/paper.bib
@@ -115,7 +115,7 @@
   journaltitle = {{IEEE} reviews in Biomedical Engineering},
   author       = {Shamout, Farah and Zhu, Tingting and Clifton, David A.},
   doi          = {10.1109/RBME.2020.3007816},
-  date         = {2020},
+  date         = {2021},
   note         = {Publisher: {IEEE}}
 }
 
