### 2023-03-24 Aarhus-Psychiatry-Research/timeseriesflattener#155 @MartinBernstorff
Sorry about that, fixed in [f26a37d](https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener/commit/f26a37d8547031fdb8a10d2ab4fc9827cee6de90).
