### 2023-03-24 #5197 @mstimberg
Thanks for the changes. I just realized that arxiv preprints also have DOIs nowadays, so you might also consider adding the https://doi.org/10.48550/arXiv.2210.12090 doi to the AutoPrognosis reference for completeness.

Apart from that, could you please:

We can then move forward with a last check of the proof, and then I'll recommend acceptance of the submission.
