+++ paper/paper.bib
@@ -116,7 +116,8 @@
   author       = {Shamout, Farah and Zhu, Tingting and Clifton, David A.},
   doi          = {10.1109/RBME.2020.3007816},
   date         = {2021},
-  note         = {Publisher: {IEEE}}
+  note         = {Publisher: {IEEE}},
+  doi.         = {10.48550/arXiv.2210.12090}
 }
 
 @article{johnson_mimic-iii_2016,
