### 2023-03-26 dd74428
fix:(deps-dev): update sphinxext-opengraph requirement
Updates the requirements on [sphinxext-opengraph](https://github.com/wpilibsuite/sphinxext-opengraph) to permit the latest version.
- [Release notes](https://github.com/wpilibsuite/sphinxext-opengraph/releases)
- [Commits](https://github.com/wpilibsuite/sphinxext-opengraph/compare/v0.7.3...v0.8.1)

---
updated-dependencies:
- dependency-name: sphinxext-opengraph
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
