+++ pyproject.toml
@@ -44,7 +44,7 @@ docs = [
     "sphinx>=5.3.0,<5.4.0",
     "furo==2022.12.7",
     "sphinx-copybutton>=0.5.1,<0.5.2",
-    "sphinxext-opengraph>=0.7.3,<0.7.5",
+    "sphinxext-opengraph>=0.7.3,<0.8.2",
     "myst-nb>=0.6.0,<1.17.0",
     "sphinx_design>=0.3.0,<0.3.1",
     "autodoc_pydantic>=1.1.0,<1.9.0",
