+++ pyproject.toml
@@ -42,7 +42,7 @@ dev = [
 
 docs = [
     "sphinx>=5.3.0,<5.4.0",
-    "furo==2022.12.7",
+    "furo==2023.3.23",
     "sphinx-copybutton>=0.5.1,<0.5.2",
     "sphinxext-opengraph>=0.7.3,<0.7.5",
     "myst-nb>=0.6.0,<1.17.0",
