+++ pyproject.toml
@@ -41,7 +41,7 @@ dev = [
 ]
 
 docs = [
-    "sphinx>=5.3.0,<5.4.0",
+    "sphinx>=5.3.0,<6.2.0",
     "furo==2023.3.23",
     "sphinx-copybutton>=0.5.1,<0.5.2",
     "sphinxext-opengraph>=0.7.3,<0.8.2",
