### 2023-03-28 1974db2
paper: don't print title twice
