### 2023-03-28 #5197 @mstimberg
Could you (and/or your coauthors) have a final look at the proof to make sure everything is correct (in particular things I cannot easily verify, e.g. author names, affiliations, funding acknowledgements…)?

Since the paper is part of the repository/archive, it would be the cleanest to also release a new version/Zenodo archive with the change if not too much hassle.
