+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <!--next-version-placeholder-->
 
+## v0.23.10 (2023-03-28)
+### Fix
+* Force new release ([`e6522d6`](https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener/commit/e6522d6af0c7d302d7326a15cdbc7c3aaee1e71c))
+
 ## v0.23.9 (2023-03-28)
 ### Fix
 * Description in .zenodo.json ([`1a4e31a`](https://github.com/Aarhus-Psychiatry-Research/timeseriesflattener/commit/1a4e31adfd7f3253a56e312444767cd2c4d6a15a))
