### 2023-03-28 #5197 @mstimberg
Oh, I just noticed one more issue that editorialbot did not catch, unfortunately: the Shamouth et al. reference currently has two DOI entries, the second needs to go into the AutoPrognosis2.0 paper instead.
