+++ paper/paper.bib
@@ -105,6 +105,7 @@
   journaltitle = {{arXiv} preprint {arXiv}:2210.12090},
   author       = {Imrie, Fergus and Cebere, Bogdan and {McKinney}, Eoin F. and van der Schaar, Mihaela},
   date         = {2022},
+  doi          = {10.48550/arXiv.2210.12090},
   file         = {Imrie et al_2022_AutoPrognosis 2.pdf:/Users/au554730/Library/Mobile Documents/com~apple~CloudDocs/zotero_storage/Imrie et al_2022_AutoPrognosis 2.pdf:application/pdf;Snapshot:/Users/au554730/Zotero/storage/NJSFPHCH/2210.html:text/html}
 }
 
@@ -116,8 +117,7 @@
   author       = {Shamout, Farah and Zhu, Tingting and Clifton, David A.},
   doi          = {10.1109/RBME.2020.3007816},
   date         = {2021},
-  note         = {Publisher: {IEEE}},
-  doi          = {10.48550/arXiv.2210.12090}
+  note         = {Publisher: {IEEE}}
 }
 
 @article{johnson_mimic-iii_2016,
